+++ xenian_bot/utils/__init__.py
@@ -3,3 +3,4 @@ from .data import *
 from .progress_bar import *
 from .telegram import *
 from .template import *
+from .telegram_files import *